feat: Dasboard Dynamic Data Change fix: Transaction Seeder only for today month

+++ database/migrations/2024_02_08_000004_create_menus_table.php
@@ -4,7 +4,8 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -13,7 +14,7 @@ return new class extends Migration {
         Schema::create('menus', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->decimal('price');
+            $table->double('price');
             $table->string('image')->nullable();
             $table->text('description');
             $table->unsignedBigInteger('type_id');
