+++ database/migrations/2024_02_08_000008_create_transaction_details_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->unsignedBigInteger('menu_id');
             $table->unsignedBigInteger('transaction_id');
             $table->integer('qty');
-            $table->decimal('unitPrice');
-            $table->decimal('subTotal');
+            $table->double('unitPrice');
+            $table->double('subTotal');
             $table->timestamps();
         });
     }
